+++ R/cache.R
@@ -10,6 +10,8 @@ isUsingCache <- function(project) {
 # We assume 'path' is the path to a DESCRIPTION file
 #' @importFrom tools md5sum
 hash <- function(path) {
+
+
   if (!file.exists(path))
     stop("No DESCRIPTION file at path '", path, "'!")
 
@@ -17,14 +19,10 @@ hash <- function(path) {
 
   DESCRIPTION <- as.data.frame(readDcf(path), stringsAsFactors = FALSE)
 
-  # If we already have a GitHub SHA1, just use that
-  if ("GithubSHA1" %in% names(DESCRIPTION))
-    return(DESCRIPTION$GithubSHA1)
-
   # TODO: Do we want the 'Built' field used for hashing? The main problem with using that is
   # it essentially makes packages installed from source un-recoverable, since they will get
   # built transiently and installed (and so that field could never be replicated).
-  fields <- c("Package", "Version", "Depends", "Imports", "Suggests", "LinkingTo")
+  fields <- c("Package", "Version", "GithubSHA1", "Depends", "Imports", "Suggests", "LinkingTo")
   sub <- DESCRIPTION[names(DESCRIPTION) %in% fields]
 
   # Handle LinkingTo specially -- we need to discover what version of packages in LinkingTo
@@ -53,11 +51,17 @@ hash <- function(path) {
             "\nThese packages are required to be installed when attempting to hash this package for caching.",
             call. = FALSE)
   }
-  linkingToHashes <- dropNull(linkingToHashes)
+
+
+  linkingToHashes <- if (length(linkingToHashes))
+    paste(
+      collapse = "",
+      sort_c(unlist(dropNull(linkingToHashes)))
+    )
 
   # Normalize for hashing and add in the linkingTo hashes as well
   ready <- normalizeForHash(sub)
-  ready <- paste0(ready, do.call(paste0, linkingToHashes))
+  ready <- paste(ready, linkingToHashes)
   tempfile <- tempfile()
   cat(ready, file = tempfile)
   result <- md5sum(tempfile)
@@ -79,33 +83,28 @@ moveInstalledPackagesToCache <- function(project = NULL) {
   if (!file.exists(cacheLibDir()))
     dir.create(cacheLibDir(), recursive = TRUE)
 
-  project <- getProjectDir(project)
-
   ## All directories within the 'lib' directory which are not symlinks are fresh
   ## and may need to be moved
   installedPkgPaths <- list.files(libDir(project), full.names = TRUE)
   if (!length(installedPkgPaths)) return(invisible())
-  cachedPkgPaths <- list.files(cacheLibDir(), full.names = TRUE)
-  cachedPkgHashes <- list.files(cachedPkgPaths)
 
   needsMove <- installedPkgPaths[sapply(installedPkgPaths, Negate(is.symlink))]
 
   ## for each package installed that is not a symlink, we migrate it to the cache
   for (package in needsMove) {
     hash <- hash(file.path(package, "DESCRIPTION"))
+    cachedPackagePath <- cacheLibDir(basename(package), hash, basename(package))
 
     ## if the package doesn't exist in the cache, copy it there
-    if (!file.exists(cacheLibDir(basename(package), hash))) {
-      dir_copy(
-        package,
-        cacheLibDir(basename(package), hash)
-      )
-    }
+    if (!file.exists(cacheLibDir(basename(package), hash)))
+      dir_copy(package, cachedPackagePath)
 
     ## replace the local package with a symlink
-    if (!is.symlink(package)) unlink(package, recursive = TRUE)
+    if (!is.symlink(package))
+      unlink(package, recursive = TRUE)
+
     symlink(
-      normalizePath(cacheLibDir(basename(package), hash)),
+      normalizePath(cachedPackagePath),
       package
     )
   }
@@ -114,20 +113,27 @@ moveInstalledPackagesToCache <- function(project = NULL) {
 
 # Pull out cached package information from the DESCRIPTION
 cachedPackages <- function(cacheDir = cacheLibDir(), fields = NULL) {
+
   pkgCachePaths <- list.files(cacheDir, full.names = TRUE)
   pkgPaths <- setNames(lapply(pkgCachePaths, function(x) {
     list.files(x, full.names = TRUE)
   }), basename(pkgCachePaths))
 
-  lapply(pkgPaths, function(hashedPath) {
-    result <- setNames(lapply(hashedPath, function(path) {
-      as.list(readDcf(file.path(path, "DESCRIPTION"), all = TRUE))
-    }), basename(hashedPath))
+  lapply(seq_along(pkgPaths), function(i) {
+
+    pkgName <- names(pkgPaths)[[i]]
+    hashedPaths <- pkgPaths[[i]]
+
+    result <- setNames(lapply(hashedPaths, function(path) {
+      as.list(readDcf(file.path(path, pkgName, "DESCRIPTION"), all = TRUE))
+    }), pkgName)
+
     if (!is.null(fields)) {
-      result[fields]
+      lapply(result, `[`, fields)
     } else {
       result
     }
+
   })
 }