+++ inst/resources/init.R
@@ -151,7 +151,7 @@ local({
     ## an 'installed from source' version
 
     ## -- InstallAgent -- ##
-    installAgent <- 'InstallAgent: packrat 0.4.3-1'
+    installAgent <- 'InstallAgent: packrat 0.4.3-25'
 
     ## -- InstallSource -- ##
     installSource <- 'InstallSource: source'